Moved request tracker to separate module
