+++ internal/handler/uncors_handler_options.go
@@ -3,7 +3,7 @@ package handler
 import (
 	"github.com/evg4b/uncors/internal/config"
 	"github.com/evg4b/uncors/internal/contracts"
-	"github.com/evg4b/uncors/internal/tui"
+	"github.com/evg4b/uncors/internal/tui/request_tracker"
 )
 
 type RequestHandlerOption = func(*RequestHandler)
@@ -44,7 +44,7 @@ func WithMockHandlerFactory(factory MockHandlerFactory) RequestHandlerOption {
 	}
 }
 
-func WithRequestTracker(tracker tui.RequestTracker) RequestHandlerOption {
+func WithRequestTracker(tracker request_tracker.RequestTracker) RequestHandlerOption {
 	return func(h *RequestHandler) {
 		h.tracker = tracker
 	}
