+++ internal/tui/request_tracker/responces.go
@@ -1,11 +1,10 @@
-package tui
+package request_tracker
 
 import (
-	"strconv"
-
 	"github.com/charmbracelet/lipgloss"
 	"github.com/evg4b/uncors/internal/helpers"
 	"github.com/evg4b/uncors/internal/tui/styles"
+	"strconv"
 )
 
 func RenderDoneRequest(request DoneRequestDefinition) string {
