+++ internal/tui/request_tracker/tracker.go
@@ -1,15 +1,14 @@
-package tui
+package request_tracker
 
 import (
 	"fmt"
+	"github.com/charmbracelet/bubbletea"
+	"github.com/evg4b/uncors/internal/contracts"
+	"github.com/evg4b/uncors/internal/helpers"
 	"net/http"
 	"sort"
 	"strings"
 	"sync"
-
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/evg4b/uncors/internal/contracts"
-	"github.com/evg4b/uncors/internal/helpers"
 )
 
 const bufferSize = 10
