+++ internal/uncors/app.go
@@ -3,6 +3,7 @@ package uncors
 import (
 	"errors"
 	"fmt"
+	"github.com/evg4b/uncors/internal/tui/request_tracker"
 	"net"
 	"net/http"
 	"strconv"
@@ -14,7 +15,6 @@ import (
 	"github.com/evg4b/uncors/internal/config"
 	"github.com/evg4b/uncors/internal/contracts"
 	"github.com/evg4b/uncors/internal/helpers"
-	"github.com/evg4b/uncors/internal/tui"
 	"github.com/spf13/afero"
 	"golang.org/x/net/context"
 )
@@ -32,7 +32,7 @@ type App struct {
 	httpsListenerMutex *sync.Mutex
 	httpsListener      net.Listener
 	cache              appCache
-	tracker            tui.RequestTracker
+	tracker            request_tracker.RequestTracker
 }
 
 const (
@@ -41,7 +41,7 @@ const (
 	shutdownTimeout   = 15 * time.Second
 )
 
-func CreateApp(fs afero.Fs, version string, tracker tui.RequestTracker) *App {
+func CreateApp(fs afero.Fs, version string, tracker request_tracker.RequestTracker) *App {
 	return &App{
 		fs:                 fs,
 		version:            version,
